feat(auth): implement logout functionality
